requirements.txt
fastapi==0.110.1
SQLAlchemy==2.0.29
PyJWT==2.7.0
bcrypt==4.1.2
pgvector==0.2.5
uvicorn==0.14.0
psycopg2-binary==2.9.9
face_recognition==1.3.0
python-multipart==0.0.9
git+https://github.com/ageitgey/face_recognition_models